feature: model product, add request products in security; change: move route params and get productsBySubCategoryId() to nav comp; fix: path in Routes
