+++ backend/WebStoreSpring/src/main/java/com/example/porfolio/webstorespring/controllers/ProductController.java
@@ -11,7 +11,7 @@ import org.springframework.web.bind.annotation.*;
 import java.util.List;
 
 @RestController
-@CrossOrigin("http://localhost:4200")
+@CrossOrigin(origins = "http://localhost:4200")
 @RequestMapping("api/v1/subcategories/")
 @RequiredArgsConstructor
 public class ProductController {
