+++ backend/WebStoreSpring/src/main/resources/application.properties
@@ -4,3 +4,10 @@ spring.jpa.hibernate.ddl-auto=update
 spring.datasource.password=Web_Store
 spring.datasource.username=web_store
 spring.jpa.properties.hibernate.dialect = org.hibernate.dialect.MySQL8Dialect
+
+spring.mail.host=smtp.gmail.com
+spring.mail.port=587
+spring.mail.username=glowackispring@gmail.com
+spring.mail.password=pastgspgwteiapyb
+spring.mail.properties.mail.smtp.auth=true
+spring.mail.properties.mail.smtp.starttls.enable=true
