+++ frontend/WebStoreAngular/src/app/models/product.ts
@@ -1,2 +1,8 @@
 export interface Product {
+  id: string;
+  name: string;
+  description: string;
+  imageUrl: string;
+  price: string;
+  type: string;
 }
