+++ frontend/WebStoreAngular/src/app/navigation/navigation.component.ts
@@ -1,6 +1,8 @@
 import { Component, OnInit } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 import { Subscription } from 'rxjs';
 import { Category } from '../models/category';
+import { Product } from '../models/product';
 import { ShopService } from '../shop.service';
 
 @Component({
@@ -11,8 +13,9 @@ import { ShopService } from '../shop.service';
 export class NavigationComponent implements OnInit {
   private categories: Category[] = [];
   private sub: Subscription;
+  subcategoryProducts: Product[] = [];
 
-  constructor(private shopService: ShopService) {
+  constructor(private route: ActivatedRoute, private shopService: ShopService) {
     this.sub = shopService.categories.subscribe((categories) => {
       this.categories = categories;
       console.log(this.categories);
@@ -23,5 +26,22 @@ export class NavigationComponent implements OnInit {
     return this.categories;
   }
 
-  ngOnInit(): void {}
+  ngOnInit() {
+    this.route.paramMap.subscribe((params) => {
+      const subcategoryId = params.get('id') as string;
+      if (subcategoryId) {
+        this.getProductsBySubcategoryId(subcategoryId);
+      }
+    });
+  }
+
+  public getProductsBySubcategoryId(subcategoryId: string) {
+    console.log('start gets products');
+    this.shopService
+      .getProductsBySubcategory(subcategoryId)
+      .subscribe((products) => {
+        this.subcategoryProducts = products;
+      });
+    console.log(this.subcategoryProducts);
+  }
 }
