+++ frontend/WebStoreAngular/src/app/products/products.component.ts
@@ -1,8 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import { ActivatedRoute } from '@angular/router';
-import { Subscription } from 'rxjs';
 import { Product } from '../models/product';
-import { ShopService } from '../shop.service';
 
 @Component({
   selector: 'app-products',
@@ -12,23 +9,6 @@ import { ShopService } from '../shop.service';
 export class ProductsComponent implements OnInit {
   private subcategoryProducts: Product[] = [];
 
-  constructor(
-    private route: ActivatedRoute,
-    private shopService: ShopService
-  ) {}
-
-  ngOnInit() {
-    this.route.paramMap.subscribe((params) => {
-      const subcategoryId = params.get('id') as string;
-      this.getProductsBySubcategoryId(subcategoryId);
-    });
-  }
-
-  getProductsBySubcategoryId(subcategoryId: string) {
-    this.shopService
-      .getProductsBySubcategory(subcategoryId)
-      .subscribe((products) => {
-        this.subcategoryProducts = products;
-      });
-  }
+  constructor() {}
+  ngOnInit(): void {}
 }
